# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateRoomBody {
  nickname: string;
}

export interface ChatRoom {
  id: number;
  nickname: string;
  createdAt: string;
}

export interface PostMessageBody {
  sender: string;
  text: string;
}

export interface ChatMessage {
  id: number;
  roomId: number;
  sender: string;
  text: string;
  createdAt: string;
}